change import directory for ProductDetailActivity and HomeFragment

+++ app/src/main/java/com/example/shoestoreapp/activity/ProductDetailActivity.kt
@@ -14,7 +14,7 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.viewpager2.widget.ViewPager2
 import com.example.shoestoreapp.R
 import com.example.shoestoreapp.adapter.ImageSliderAdapter
-import com.example.shoestoreapp.classes.Product
+import com.example.shoestoreapp.data.model.Product
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.tbuonomo.viewpagerdotsindicator.DotsIndicator
@@ -70,7 +70,7 @@ class ProductDetailActivity : AppCompatActivity() {
         productTitleTV.text = product.name
         priceTV.text = product.price?.toString() ?: "0"
         productDescriptionTV.text = product.description
-        ratingStar.rating = product.averageRating ?: 0f
+        ratingStar.rating = (product.averageRating ?: 0f).toFloat()
 
         // Cài đặt ViewPager2 với Adapter
         val adapter = ImageSliderAdapter(product.images ?: emptyList())
